Remove PathIterator trait. paths_iter method is now provided by the Fst trait
