+++ rustfst-python/rustfst/iterators.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 import ctypes
 from typing import Optional
+
+from rustfst import Fst
 from rustfst.utils import lib, check_ffi_error
 from rustfst.tr import Tr
 
@@ -11,7 +13,7 @@ class TrsIterator:
       This class is used for iterating over the trs leaving some state of a FST.
     """
 
-    def __init__(self, fst: Fst, state: int) -> TrsIterator:
+    def __init__(self, fst: Fst, state: int):
         self.ptr = fst  # reference fst to prolong its lifetime (prevent early gc)
         state = ctypes.c_size_t(state)
         iter_ptr = ctypes.pointer(ctypes.c_void_p())
