+++ rustfst/src/algorithms/union/union.rs
@@ -19,7 +19,7 @@ use crate::{StateId, Trs, EPS_LABEL};
 /// # use rustfst::utils::transducer;
 /// # use rustfst::semirings::{Semiring, IntegerWeight};
 /// # use rustfst::fst_impls::VectorFst;
-/// # use rustfst::fst_traits::PathsIterator;
+/// # use rustfst::fst_traits::Fst;
 /// # use rustfst::FstPath;
 /// # use rustfst::algorithms::union::union;
 /// # use std::collections::HashSet;
