+++ rustfst/src/fst_traits/mod.rs
@@ -3,7 +3,6 @@ pub use self::expanded_fst::ExpandedFst;
 pub use self::fst::{CoreFst, Fst};
 pub use self::iterators::{FstIntoIterator, FstIterData, FstIterator, StateIterator};
 pub use self::mutable_fst::MutableFst;
-pub use self::paths_iterator::PathsIterator;
 pub use self::serializable_fst::SerializableFst;
 
 #[macro_use]
@@ -16,3 +15,4 @@ mod iterators;
 mod mutable_fst;
 mod paths_iterator;
 mod serializable_fst;
+mod string_paths_iterator;
